+++ spec/www/spec/db-open-close-delete-test.js
@@ -320,8 +320,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'Open database with U+0801 (3-byte Samaritan character Bit) & check internal database file name', function(done) {
-          if (!isWindows && isAndroid && !isImpl2) pending('XXX CRASH on Android (default evcore-native-driver database access implementation)');
-
           var dbName = 'a\u0801.db';
 
           try {
@@ -360,7 +358,7 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'Open database with emoji \uD83D\uDE03 (UTF-8 4 bytes) & check internal database file name', function(done) {
-          if (!isWindows && isAndroid && !isImpl2) pending('XXX CRASH on Android (default evcore-native-driver database access implementation)');
+          if (!isWindows && isAndroid && !isImpl2) pending('TBD KNOWN ISSUE on Android (default evcore-native-driver database access implementation)');
 
           var dbName = 'a\uD83D\uDE03.db';