+++ spec/www/spec/db-tx-value-bindings-test.js
@@ -531,17 +531,13 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        // NOTE: emojis and other 4-octet UTF-8 characters _evidently_ not stored
-        // properly by Android-sqlite-connector / Android evcore-native-driver
-        // XXX WITH KNOWN CRASH ISSUE on Android evcore-native-driver
+        // TBD ENCODING ISSUE iwth emojis and other 4-octet UTF-8 characters
+        // on Android (default Android evcore NDK implementation)
         // ref:
+        // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
         // - litehelpers/Cordova-sqlite-storage#564
-        // - litehelpers/Cordova-sqlite-evcore-extbuild-free#19
-        // - litehelpers/Android-sqlite-evcore-native-driver-free#1
-        // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7 (XXX KNOWN CRASH issue on Android)
-        it(suiteName + 'INSERT TEXT string with emoji [\\u1F603 SMILING FACE (MOUTH OPEN)], SELECT the data, check, and check HEX [XXX TBD TRUNCATION BUG REPRODUCED on Windows; default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]' , function(done) {
-          if (!isWebSql && !isWindows && isAndroid && !isImpl2) pending('XXX KNOWN CRASH ISSUE on Android (default evcore-native-driver database access implementation)'); // XXX
-
+        // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7
+        it(suiteName + 'INSERT TEXT string with emoji [\\u1F603 SMILING FACE (MOUTH OPEN)], SELECT the data, check, and check HEX [ENCODING ISSUE REPRODUCED on Android post-5.x; non-standard encoding reproduced on Android pre-6.0; TRUNCATION BUG REPRODUCED on Windows; default sqlite HEX encoding: UTF-6le on Windows & Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]' , function(done) {
           var db = openDatabase('INSERT-emoji-and-check.db', '1.0', 'Demo', DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -559,27 +555,27 @@ var mytests = function() {
 
                   var row = rs2.rows.item(0);
                   // Full object check:
-                  expect(row).toEqual({data: '@\uD83D\uDE03!'});
+                  if (!isWebSql && !isWindows && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
+                    expect(row).toEqual({data: '@?!'});
+                  else
+                    expect(row).toEqual({data: '@\uD83D\uDE03!'});
                   // Check individual members:
-                  expect(row.data).toBe('@\uD83D\uDE03!');
+                  if (!isWebSql && !isWindows && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
+                    expect(row.data).toBe('@?!');
+                  else
+                    expect(row.data).toBe('@\uD83D\uDE03!');
 
                   tx.executeSql('SELECT HEX(data) AS hexvalue FROM test_table', [], function(tx_ignored, rs3) {
                     expect(rs3).toBeDefined();
                     expect(rs3.rows).toBeDefined();
                     expect(rs3.rows.length).toBe(1);
 
-                    // TBD NOT APPLICABLE in this plugin version
-                    // (TEST SKIPPPED for default evcore-native-driver database
-                    //  access implementation due to possible crash on
-                    //  Android 7.x/???):
-                    // STOP HERE [HEX encoding BUG] for XXX TBD
-                    // if (!isWebSql && !isWindows && isAndroid && !isImpl2) return done();
-
                     if (isWindows || (isWebSql && isAndroid && /Android 4.[1-3]/.test(navigator.userAgent)))
                       expect(rs3.rows.item(0).hexvalue).toBe('40003DD803DE2100'); // (UTF-16le)
-                    /* XXX TBD HEX encoding BUG IGNORED on default Android NDK access implementation (...)
+                    else if (!isWebSql && !isWindows && isAndroid && !isImpl2 && /Android [4-5]/.test(navigator.userAgent))
+                      expect(rs3.rows.item(0).hexvalue).toBe('40EDA0BDEDB88321'); // non-standard encoding on Android pre-6.0
                     else if (!isWebSql && isAndroid && !isImpl2)
-                      expect(rs3.rows.item(0).hexvalue).toBe('--'); // (...) */
+                      expect(rs3.rows.item(0).hexvalue).toBe('403F21'); // ENCODING ISSUE on Anroid post-5.x UTF-8
                     else
                       expect(rs3.rows.item(0).hexvalue).toBe('40F09F988321'); // (UTF-8)
 
@@ -599,9 +595,12 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT TEXT string with 25 emojis, SELECT the data, and check' , function(done) {
-          if (!isWebSql && !isWindows && isAndroid && !isImpl2) pending('XXX KNOWN CRASH ISSUE on Android (default evcore-native-driver database access implementation)'); // XXX
-
+        it(suiteName + 'INSERT TEXT string with 25 emojis, SELECT the data, and check - ENCODING ISSUE REPRODUCED on Android post-5.x (default evcore NDK implementation)' , function(done) {
+          // ref:
+          // - litehelpers/Cordova-sqlite-evcore-extbuild-free#44
+          // - litehelpers/Cordova-sqlite-storage#564
+          // - litehelpers/Cordova-sqlite-evcore-extbuild-free#43
+          // - litehelpers/Cordova-sqlite-evcore-extbuild-free#7
           var db = openDatabase('INSERT-emoji-and-check.db', '1.0', 'Demo', DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -619,7 +618,10 @@ var mytests = function() {
                   expect(rs2.rows.length).toBe(1);
 
                   var row = rs2.rows.item(0);
-                  expect(row.data).toBe(part + part + part + part + part);
+                  if (!isWebSql && isAndroid && !isImpl2 && !(/Android [4-5]/.test(navigator.userAgent)))
+                    expect(row.data).toBe('@?????@?????@?????@?????@?????');
+                  else
+                    expect(row.data).toBe(part + part + part + part + part);
 
                   // Close (plugin only) & finish:
                   (isWebSql) ? done() : db.close(done, done);
@@ -1091,6 +1093,7 @@ var mytests = function() {
       describe(suiteName + 'INLINE BLOB value storage tests', function() {
 
         it(suiteName + "INSERT inline BLOB value (X'40414243') and check stored data [TBD SELECT BLOB value ERROR EXPECTED on Windows and Android with androidDatabaseImplementation: 2 setting; with default sqlite HEX encoding: UTF-6le on Android 4.1-4.3 (WebKit) Web SQL, UTF-8 otherwise]", function(done) {
+          // ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#43
           var db = openDatabase('INSERT-inline-BLOB-value-40414243-and-check-stored-data.db');
 
           db.transaction(function(tx) {
@@ -1159,7 +1162,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + "INSERT inline BLOB value (X'FFD1FFD2') and check stored data [XXX SKIP FINAL SELECT CHECK on default Android NDK access implementation due to CRASH ISSUE; OTHER PLUGIN ISSUES REPRODUCED: missing result column data; SELECT BLOB VALUE ERROR on Android (androidDatabaseProvider: 'system') & Windows; missing result column data on iOS/macOS]", function(done) {
+        it(suiteName + "INSERT inline BLOB value (X'FFD1FFD2') and check stored data [PLUGIN ISSUES REPRODUCED: SELECT BLOB returns data with incorrect length on Android (default NDK implementation); SELECT BLOB VALUE ERROR on Android (androidDatabaseProvider: 'system') & Windows; missing result column data on iOS/macOS]", function(done) {
           var db = openDatabase('INSERT-inline-BLOB-value-FFD1FFD2-and-check-stored-data.db', '1.0', 'Demo', DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -1180,9 +1183,6 @@ var mytests = function() {
                   expect(item).toBeDefined();
                   expect(item.hexValue).toBe('FFD1FFD2');
 
-                  // STOP HERE to avoid CRASH on default Android NDK access implementation:
-                  if (!isWebSql && !isWindows && isAndroid && !isImpl2) return done();
-
                   tx.executeSql('SELECT * FROM test_table', [], function(ignored, rs3) {
                     if (!isWebSql && isWindows) expect('PLUGIN BEHAVIOR CHANGED: UNEXPECTED SUCCESS on Windows').toBe('--');
                     if (!isWebSql && isAndroid && isImpl2) expect('PLUGIN BEHAVIOR CHANGED: UNEXPECTED SUCCESS on Android with androidDatabaseImplementation: 2 setting').toBe('--');
@@ -1202,6 +1202,9 @@ var mytests = function() {
                       // EXPECTED RESULT on (WebKit) Web SQL & FUTURE TBD: plugin on browser platform
                       expect(mydata).toBeDefined();
                       expect(mydata.length).toBeDefined();
+                      if (!isWebSql && isAndroid && !isImpl2)
+                        expect(mydata.length).toBe(1); // INCORRECT DATA LENGTH on Android (default evcore NDK)
+                      else
                       if (!(isWebSql && /Android 4.[1-3]/.test(navigator.userAgent)))
                         expect(mydata.length).toBe(4);
                     }