+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
-### cordova-sqlite-evcore-common-free 0.0.4-pre2
+### cordova-sqlite-evcore-common-free 0.0.4
+
+- Use cordova-sqlite-evcore-free-dependencies 0.8.6, with workaround for 4-byte UTF-8 crash bug and fix for Samaritan character crash bug on Android
 
 #### cordova-sqlite-storage 2.5.1